builds out options component
